+++ launcher/src/main/java/io/github/bekoenig/getdown/launcher/ProxyPanel.java
@@ -23,9 +23,9 @@ import javax.swing.JPanel;
 import javax.swing.JPasswordField;
 import javax.swing.JTextField;
 
-import com.samskivert.swing.GroupLayout;
-import com.samskivert.swing.Spacer;
-import com.samskivert.swing.VGroupLayout;
+import io.github.bekoenig.getdown.launcher.swing.GroupLayout;
+import io.github.bekoenig.getdown.launcher.swing.Spacer;
+import io.github.bekoenig.getdown.launcher.swing.VGroupLayout;
 import io.github.bekoenig.getdown.util.MessageUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
