+++ launcher/src/main/java/io/github/bekoenig/getdown/launcher/swing/util/Throttle.java
@@ -18,7 +18,7 @@
 // License along with this library; if not, write to the Free Software
 // Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
 
-package com.samskivert.util;
+package io.github.bekoenig.getdown.launcher.swing.util;
 
 /**
  * A throttle is used to prevent code from attempting a particular operation too often. Often it is
